+++ hydra-node/src/Hydra/Node/Run.hs
@@ -98,7 +98,7 @@ run opts = do
             =<< createPersistenceIncremental stateFile
         -- NOTE: Add any custom sinks here
         let eventSinks :: [EventSink (StateEvent Tx) IO] = []
-        wetHydraNode <- hydrate (contramap Node tracer) env ledger initialChainState eventStore eventSinks
+        wetHydraNode <- hydrate (contramap Node tracer) env ledger initialChainState persistenceDir eventStore eventSinks
         -- Chain
         withChain <- prepareChainComponent tracer env chainConfig
         withChain (chainStateHistory wetHydraNode) (wireChainInput wetHydraNode) $ \chain -> do