+++ hydra-tx/test/Hydra/Tx/HeadParametersSpec.hs
@@ -1,0 +1,14 @@
+module Hydra.Tx.HeadParametersSpec where
+
+import Hydra.Prelude
+import Test.Hydra.Prelude
+
+import Hydra.Tx.HeadParameters (HeadParameters)
+import Hydra.Tx.IsTxSpec (roundtripCBOR)
+
+spec :: Spec
+spec =
+  parallel $ do
+    describe "HeadParameters (cardano)" $
+      prop "Roundtrip CBOR encoding HeadParameters" $
+        roundtripCBOR @HeadParameters